+++ src/utils/externalSiteMap.ts
@@ -1,0 +1,6 @@
+export enum ExternalSiteMap {
+	CuriousCrittersSong = "https://opengameart.org/content/curious-critters",
+	RetroRpgMenuSounds = "https://opengameart.org/content/10-retro-rpg-menu-sounds",
+	CreativeCommonsV3 = "https://creativecommons.org/licenses/by/3.0/",
+	CreativeCommonsV4 = "https://creativecommons.org/licenses/by/4.0/",
+}