+++ src/app/contact/components/ffContactForm.tsx
@@ -15,18 +15,18 @@ export default function FfContactForm() {
 
 	
 	return (
-		<div  className="flex flex-col h-[99vh] w-[99vw] gap-2 pt-2 m-auto justify-around bg-black">
+		<div  className="flex flex-col h-[99vh] w-[99vw] gap-1 pt-2 m-auto justify-around bg-black">
 			<audio ref={audioRef} autoPlay loop src={ MediaMap.CURIOUS_CRITTERS }></audio>
 
 			<RadioInputForm contactReason={contactReason} onChange={setContactReason}/>
 			
-			<Form type="submit" action={emailUserMessage} className="flex flex-col gap-2"> 
-				<div className="flex flex-1 gap-2">
+			<Form type="submit" action={emailUserMessage} className="flex flex-col gap-1"> 
+				<div className="flex flex-1 gap-1">
 					<TextInputForm name="name" inputPrompt="What is your name?" />
 					<TextInputForm name="email" inputPrompt="What is your email address?" />
 				</div>
 
-				<div className="flex flex-1 gap-2">
+				<div className="flex flex-1 gap-1">
 					<input value={contactReason} name="reason" hidden readOnly/>
 					<TextInputForm name="message" inputPrompt="Leave me a message and I'll get back to you" />
 					<AudioCredits/>