+++ src/app/contact/formProcess/page.tsx
@@ -1,0 +1,5 @@
+import MessageSuccess from "./messageSuccess";
+
+export default function FormProcess() {
+	return <MessageSuccess/>
+}